reworked redux work to parent to prevent loading issues
