+++ client/app/queue/correspondence/CorrespondenceCases.jsx
@@ -1,5 +1,8 @@
 import * as React from 'react';
-
+import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
+import ApiUtil from '../../util/ApiUtil';
+import { loadVetCorrespondence } from './correspondenceReducer/correspondenceActions';
 import AppSegment from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/AppSegment';
 import PropTypes from 'prop-types';
 import COPY from '../../../COPY';
@@ -13,6 +16,26 @@ import QueueOrganizationDropdown from '../components/QueueOrganizationDropdown';
 // } from '../QueueActions';
 
 class CorrespondenceCases extends React.PureComponent {
+
+  // grabs correspondences and loads into intakeCorrespondence redux store.
+  getVeteransWithCorrespondence() {
+    return ApiUtil.get('/queue/correspondence?json').then((response) => {
+      const returnedObject = response.body;
+      const vetCorrespondences = returnedObject.vetCorrespondences;
+
+      this.props.loadVetCorrespondence(vetCorrespondences);
+    }).
+      catch((err) => {
+        // allow HTTP errors to fall on the floor via the console.
+        console.error(new Error(`Problem with GET /queue/correspondence?json ${err}`));
+      });
+  }
+
+  // load veteran correspondence info on page load
+  componentDidMount() {
+    this.getVeteransWithCorrespondence();
+  }
+
   render = () => {
     const {
       organizations
@@ -23,7 +46,11 @@ class CorrespondenceCases extends React.PureComponent {
         <AppSegment filledBackground>
           <h1 {...css({ display: 'inline-block' })}>{COPY.CASE_LIST_TABLE_QUEUE_DROPDOWN_CORRESPONDENCE_CASES}</h1>
           <QueueOrganizationDropdown organizations={organizations} />
-          <CorrespondenceTable />
+          {this.props.vetCorrespondences &&
+          <CorrespondenceTable
+            vetCorrespondences={this.props.vetCorrespondences}
+          />
+          }
         </AppSegment>
       </React.Fragment>
     );
@@ -31,7 +58,22 @@ class CorrespondenceCases extends React.PureComponent {
 }
 
 CorrespondenceCases.propTypes = {
-  organizations: PropTypes.array
+  organizations: PropTypes.array,
+  loadVetCorrespondence: PropTypes.func,
+  vetCorrespondences: PropTypes.array
 };
 
-export default CorrespondenceCases;
+const mapStateToProps = (state) => ({
+  vetCorrespondences: state.intakeCorrespondence.vetCorrespondences
+});
+
+const mapDispatchToProps = (dispatch) => (
+  bindActionCreators({
+    loadVetCorrespondence
+  }, dispatch)
+);
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(CorrespondenceCases);
