+++ client/app/queue/correspondence/CorrespondenceTable.jsx
@@ -1,46 +1,11 @@
 import React from 'react';
 import QueueTable from '../QueueTable';
 import PropTypes from 'prop-types';
-import ApiUtil from '../../../app/util/ApiUtil';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
-import { loadVetCorrespondence } from './correspondenceReducer/correspondenceActions';
 import Link from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/Link';
 
 class CorrespondenceTable extends React.Component {
 
-  // grabs correspondences and loads into intakeCorrespondence redux store.
-  getVeteransWithCorrespondence() {
-    return ApiUtil.get('/queue/correspondence?json').then((response) => {
-      const returnedObject = response.body;
-      const vetCorrespondences = returnedObject.vetCorrespondences;
-
-      this.props.loadVetCorrespondence(vetCorrespondences);
-    }).
-      catch((err) => {
-        // allow HTTP errors to fall on the floor via the console.
-        console.error(new Error(`Problem with GET /queue/correspondence?json ${err}`));
-      });
-  }
-
-  // load veteran correspondence info on page load
-  componentDidMount() {
-    this.getVeteransWithCorrespondence();
-  }
   render() {
-    // test data names link to multi_correspondence.rb seed data
-    const testobj = [{
-      veteranDetails: 'Adam West (66555444)',
-      packageDocumentType: '1111111',
-      cmPacketNumber: '22222222',
-      correspondeceId: 11,
-    },
-    {
-      veteranDetails: 'Michael Keaton (67555444)',
-      packageDocumentType: '12345678',
-      cmPacketNumber: '12345679',
-      correspondeceId: 2,
-    }];
 
     const columns = [
       {
@@ -82,7 +47,7 @@ class CorrespondenceTable extends React.Component {
       <QueueTable
         className="assign-correspondence-table"
         columns={columns}
-        rowObjects={testobj}
+        rowObjects={this.props.vetCorrespondences}
         summary="scheduled-hearings-table"
         enablePagination
         tabPaginationOptions={tabPaginationOptions}
@@ -103,17 +68,4 @@ CorrespondenceTable.propTypes = {
   })
 };
 
-const mapStateToProps = (state) => ({
-  vetCorrespondences: state.intakeCorrespondence.vetCorrespondences
-});
-
-const mapDispatchToProps = (dispatch) => (
-  bindActionCreators({
-    loadVetCorrespondence
-  }, dispatch)
-);
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(CorrespondenceTable);
+export default CorrespondenceTable;
